refactor localStorage related code
